+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Unreleased
 
+## Added
+- none-ls: use configuration file for markdown lint cli
+
 # 2025-07-03
 
 ## Added